mapping_write: the GPU copy into a mapping's pages moves below the CPU one
Thirteen `cfg(feature = "backend-vulkan")` items sat interleaved with the
neutral write. The neutral half is the guest's own declaration — where a
mapping's plane is, what extent and format a write into it must use, and
the copy of bytes a caller already holds — and it reads the same on any
rail. The gated half is the GPU copy, Vulkan's alone because what it copies
is: `write_bgra8_from_resident_gpu` takes a `TargetIdentity`, and
`licence_type11_surface` hands back a `GuestPageTarget`, which is a
permission only a rail holding an import of the guest's pages can grant.

They become `mapping_write/vulkan.rs`, gated once on the `mod`. The nine
rail-gated tests go with them rather than staying in the neutral parent's
test file: two of them exercise `plan_guest_window` and
`guest_store_needs_separate_sync`, which are private to the rail, and
keeping them where they were would have meant widening those to `pub(super)`
for the test's benefit alone.

Pure placement; no item's body changed. metal 1397, vulkan 2181, both 2265,
all unchanged; matrix, fmt and clippy clean, and the rustdoc warning count
is where it was.
