+++ crates/reims-vgpu/src/runtime/blit_exec/vulkan.rs
@@ -143,7 +143,7 @@ pub(crate) fn try_copy_whole_plane_on_gpu<M: HostMemory + HostOps>(
         return None;
     }
     let identity = crate::runtime::writeback_debt::gva_identity(debt);
-    match mapping_write::write_bgra8_from_resident_gpu(
+    match mapping_write::vulkan::write_bgra8_from_resident_gpu(
         state,
         host,
         mapping_id,
