+++ crates/reims-vgpu/src/runtime/mapping_write/mod.rs
@@ -1,5 +1,10 @@
 //! Write host BGRA8 into a guest IOSurface mapping (render writeback).
 //!
+//! This half is the CPU write and the geometry every rail resolves the same
+//! way: where a mapping's plane is, what extent and format a write into it must
+//! use, and the bytes-in-hand copy. [`vulkan`] is the other half — the same
+//! frame landed from a resident, without the bytes ever existing on the host.
+//!
 //! Product writes go **only** through a revalidated contiguous HostOps view
 //! (`map_pages`) — never `write_gpa` fragment walks over cached PFNs (freelist
 //! `0xff000000ff000000` class). Always bumps [`DeviceState::mark_mapping_written`]
@@ -12,6 +17,14 @@ use crate::contract::pixel_format::{
 use crate::model::{scanout_extent_ok, DeviceState, MappingEntry, MAX_SCANOUT_DIM};
 use crate::runtime::host::{HostMemory, HostOps};
 use crate::runtime::mapper;
+
+/// The rail that can copy a resident straight into a mapping's guest pages.
+///
+/// Gated on the build — whether this binary carries a Vulkan rail at all is a
+/// fact about the build, and every signature in there names a `TargetIdentity`
+/// or a `GuestPageTarget`. *Which rail is running* is never asked here.
+#[cfg(feature = "backend-vulkan")]
+pub mod vulkan;
 
 /// Why one render writeback did not land its frame in the guest's pages.
 ///
@@ -630,255 +643,7 @@ pub fn write_bgra8_uncached<M: HostMemory + HostOps>(
     )
 }
 
-/// A check that stopped a resident's frame from reaching the guest's pages
-/// without a host copy, so the flush owes the copying rail instead.
-///
-/// Every variant is a routing answer and not a loss — the caller still lands the
-/// frame — but each one is a whole frame's worth of memcpy the device paid twice
-/// over, on the rail that is 69% of the drain worker's time. So they are named
-/// individually: "the GPU writeback declined" cannot tell a host whose GPU
-/// cannot import guest RAM from a surface whose row pitch is not a whole texel,
-/// and those have different fixes.
-#[cfg(feature = "backend-vulkan")]
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum GpuWritebackDecline {
-    /// A zero or over-large rect, or a mapping that is gone or unmapped. The
-    /// copying rail refuses these too, by its own [`SurfaceWriteRefusal`]; this
-    /// path declines before making the guest pay two refusals for one cause.
-    NotWritable,
-    /// The mapping's declared geometry is not the frame's. The copying rail
-    /// reports this as `GeometryMoved`; here it means the same thing and the
-    /// same rail will say so.
-    GeometryMoved {
-        latched_width: u32,
-        latched_height: u32,
-        frame_width: u32,
-        frame_height: u32,
-    },
-    /// No sample window resolves for this geometry, so there is no destination
-    /// offset or row pitch to copy against.
-    WindowUnresolved {
-        width: u32,
-        height: u32,
-        format: u16,
-    },
-    /// The mapping's pixel format is not the one the resident holds, so landing
-    /// it needs a per-row conversion. A buffer→image copy performs none, which
-    /// is why this is a routing answer rather than something to work around.
-    FormatNeedsConversion { format: u16 },
-    /// The mapping's declared format has a linear texel to name, and the source
-    /// image does not hold it. Same consequence as the variant above and a
-    /// different cause: there the guest declared something no copy can express,
-    /// here the two sides simply disagree.
-    ///
-    /// Expected on the compute rail, where a storage image's format comes from
-    /// the specialized SPIR-V texel format and owes a surface mapping's
-    /// declaration nothing. Whole formats, not channel orders: `RGBA16_FLOAT`
-    /// and `RGBA8_UNORM` are both RGBA-ordered and four bytes per texel apart,
-    /// so an order comparison would admit a half-float source over an eight-bit
-    /// destination and land a frame of the wrong size.
-    ResidentFormatMismatch {
-        held: ash::vk::Format,
-        want: ash::vk::Format,
-    },
-    /// The guest's row pitch is not a whole number of texels, so it cannot be
-    /// expressed as `bufferRowLength`.
-    PitchNotTexels { bpr: u32 },
-    /// The frame's first texel does not start on a 4-byte boundary within its
-    /// page. `VkBufferImageCopy::bufferOffset` must be a multiple of the texel
-    /// block size, and a copy that ignored this is undefined rather than
-    /// misaligned.
-    OffsetNotTexelAligned { in_page: u64 },
-    /// The mapping's page list does not cover the sample window.
-    PageListShort { need: usize, have: usize },
-    /// A page in the window carries no valid entry, so there is no guest page
-    /// to resolve a reference against.
-    PageUnbacked { index: usize },
-    /// The page walk refused: these are no longer provably the mapping's pages.
-    /// The copying rail refuses for the same reason and reports it.
-    PagesNotOurs,
-    /// This window's pages did not become a reference. Carries the check
-    /// [`crate::runtime::guest_ram_map`] refused on — including the one that
-    /// says nothing about the window at all, that this host cannot import guest
-    /// RAM. There is deliberately no separate variant for that: the early-out
-    /// above and the walk below now ask one function, so they cannot name the
-    /// same host two ways, and `via=` says which check answered either way.
-    ///
-    /// It carries it rather than pointing at it. That module reports each
-    /// distinct refusal **once per boot** — `report_once` latches on
-    /// `first_sight` — while this decline is reached once per flush, so a boot
-    /// where every 1080p writeback is refused prints twenty of these against a
-    /// single `guest_ram_map` line elsewhere in the log. Nothing relates the
-    /// two, and the reader's obvious move — ranking `reason=` on the fail
-    /// channel — puts the twenty at the top under a name that says the host
-    /// cannot import, on a host whose `vk_caps` says `supported`. Restating the
-    /// inner check is the cheaper error.
-    GuestRefRefused {
-        refusal: crate::runtime::guest_ram_map::MapRefusal,
-    },
-    /// The engine declined or the copy failed; the inner error names which.
-    Engine {
-        inner: crate::backend::vulkan::engine::DrawError,
-    },
-}
-
-#[cfg(feature = "backend-vulkan")]
-impl crate::observe::Decline for GpuWritebackDecline {
-    fn slug(&self) -> &'static str {
-        match self {
-            Self::NotWritable => "gpuwb_not_writable",
-            Self::GeometryMoved { .. } => "gpuwb_geometry_moved",
-            Self::WindowUnresolved { .. } => "gpuwb_window_unresolved",
-            Self::FormatNeedsConversion { .. } => "gpuwb_format_needs_conversion",
-            Self::ResidentFormatMismatch { .. } => "gpuwb_resident_format_mismatch",
-            Self::PitchNotTexels { .. } => "gpuwb_pitch_not_texels",
-            Self::OffsetNotTexelAligned { .. } => "gpuwb_offset_not_texel_aligned",
-            Self::PageListShort { .. } => "gpuwb_page_list_short",
-            Self::PageUnbacked { .. } => "gpuwb_page_unbacked",
-            Self::PagesNotOurs => "gpuwb_pages_not_ours",
-            Self::GuestRefRefused { .. } => "gpuwb_guest_ref_refused",
-            // The engine's own slug, so a driver that refuses the pointer and a
-            // resident in the wrong channel order stay as distinguishable here
-            // as they are where they were decided.
-            Self::Engine { inner } => crate::observe::Decline::slug(inner),
-        }
-    }
-
-    fn fields(&self) -> Vec<(&'static str, String)> {
-        match self {
-            Self::NotWritable | Self::PagesNotOurs => Vec::new(),
-            // `via` before the inner fields, so the check that refused reads
-            // first and its own `pages=` / `first=` qualify it rather than
-            // looking like this rail's own numbers.
-            Self::GuestRefRefused { refusal } => {
-                let mut f = vec![("via", crate::observe::Decline::slug(refusal).to_string())];
-                f.extend(crate::observe::Decline::fields(refusal));
-                f
-            }
-            Self::GeometryMoved {
-                latched_width,
-                latched_height,
-                frame_width,
-                frame_height,
-            } => vec![
-                ("latched", format!("{latched_width}x{latched_height}")),
-                ("frame", format!("{frame_width}x{frame_height}")),
-            ],
-            Self::WindowUnresolved {
-                width,
-                height,
-                format,
-            } => vec![
-                ("geom", format!("{width}x{height}")),
-                ("fmt", format!("{format:#x}")),
-            ],
-            Self::FormatNeedsConversion { format } => vec![("fmt", format!("{format:#x}"))],
-            Self::ResidentFormatMismatch { held, want } => {
-                vec![("held", format!("{held:?}")), ("want", format!("{want:?}"))]
-            }
-            Self::PitchNotTexels { bpr } => vec![("bpr", bpr.to_string())],
-            Self::OffsetNotTexelAligned { in_page } => vec![("in_page", in_page.to_string())],
-            Self::PageListShort { need, have } => {
-                vec![("need", need.to_string()), ("have", have.to_string())]
-            }
-            Self::PageUnbacked { index } => vec![("page", index.to_string())],
-            Self::Engine { inner } => crate::observe::Decline::fields(inner),
-        }
-    }
-}
-
-#[cfg(feature = "backend-vulkan")]
-crate::observe::decline::decline_display!(GpuWritebackDecline);
-
-/// Which of a mapping's pages a writeback's texels live in, and where inside
-/// them the first one is.
-///
-/// The guest reference this rail binds names whole pages and starts at a page
-/// boundary; a sample window starts wherever the guest's plane descriptor put
-/// it. This is the translation between the two, and getting it wrong lands a
-/// frame at the wrong offset in the guest's memory — which is a visibly shifted
-/// surface at best and another allocation's bytes at worst.
-#[cfg(feature = "backend-vulkan")]
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-struct GuestWindowPlan {
-    /// Indices into `page_entries` of the first and last page the frame touches.
-    first_page: usize,
-    last_page: usize,
-    /// Byte offset of the frame's first texel within page `first_page`, which is
-    /// therefore its offset within the guest reference the copy binds.
-    in_page: u64,
-    /// Guest row pitch in texels (`bufferRowLength`).
-    row_length_texels: u32,
-}
-
-#[cfg(feature = "backend-vulkan")]
-impl GuestWindowPlan {
-    fn pages(&self) -> usize {
-        self.last_page - self.first_page + 1
-    }
-}
-
-/// Resolve a sample window against a mapping's page list.
-///
-/// Pure, and separate from its one caller for that reason: every value it
-/// produces feeds a `VkBufferImageCopy` whose failure mode is silent — Vulkan
-/// will happily write a frame at the wrong offset — and none of the surrounding
-/// device state is needed to decide any of them.
-#[cfg(feature = "backend-vulkan")]
-fn plan_guest_window(
-    page_entries: usize,
-    page_size: u64,
-    base_off: u64,
-    span_end: u64,
-    bpr: u32,
-    width: u32,
-    texel: u32,
-) -> Result<GuestWindowPlan, GpuWritebackDecline> {
-    // `bufferRowLength` is in texels, so a pitch that is not a whole number of
-    // them has no spelling. Checked rather than assumed: the value comes from
-    // the guest's own device descriptor.
-    //
-    // In *this destination's* texels and not in four bytes. A half-float plane
-    // states its pitch in eight-byte texels, so dividing by four would report
-    // twice as many of them — a `bufferRowLength` that reads as a valid padded
-    // stride and lands every row at half its true spacing.
-    //
-    // The second half is a Vulkan validity rule rather than an arithmetic one:
-    // `bufferRowLength` must be zero or at least `imageExtent.width`, so a pitch
-    // narrower than the frame is an invalid copy and not a tight one. It cannot
-    // happen for a well-formed plane, which is exactly why nothing would notice
-    // if it did.
-    if texel == 0 || !bpr.is_multiple_of(texel) || bpr / texel < width {
-        return Err(GpuWritebackDecline::PitchNotTexels { bpr });
-    }
-    if span_end <= base_off || page_size == 0 {
-        return Err(GpuWritebackDecline::NotWritable);
-    }
-    let first_page = (base_off / page_size) as usize;
-    let last_page = ((span_end - 1) / page_size) as usize;
-    if page_entries <= last_page {
-        return Err(GpuWritebackDecline::PageListShort {
-            need: last_page + 1,
-            have: page_entries,
-        });
-    }
-    // The guest reference starts at a page boundary, so the frame's first texel
-    // sits this far into it. Whole texels only, which is what `bufferOffset`
-    // requires and what a guest pitch in texels already implies for every row
-    // but the first.
-    let in_page = base_off % page_size;
-    if !in_page.is_multiple_of(u64::from(texel)) {
-        return Err(GpuWritebackDecline::OffsetNotTexelAligned { in_page });
-    }
-    Ok(GuestWindowPlan {
-        first_page,
-        last_page,
-        in_page,
-        row_length_texels: bpr / texel,
-    })
-}
-
-/// The plane of `m` that [`write_bgra8_from_resident_gpu`] would write a frame
+/// The plane of `m` that [`vulkan::write_bgra8_from_resident_gpu`] would write a frame
 /// of this extent into, as `(surface_offset, row_stride, pixel_format)`.
 ///
 /// # Why a caller has to ask
@@ -894,7 +659,7 @@ fn plan_guest_window(
 ///
 /// `None` means the rail would decline, so a caller that gets it owes the frame
 /// to whatever path it was going to take anyway. The rail names *which* check
-/// declined, through [`GpuWritebackDecline`]; this collapses them, because a
+/// declined, through [`vulkan::GpuWritebackDecline`]; this collapses them, because a
 /// pre-question only needs to know that the answer is not "yes".
 pub fn resident_gpu_plane(m: &MappingEntry, width: u32, height: u32) -> Option<(u64, u32, u16)> {
     let (mw, mh, format) = mapping_write_geometry(m, width, height);
@@ -903,504 +668,6 @@ pub fn resident_gpu_plane(m: &MappingEntry, width: u32, height: u32) -> Option<(
     }
     let (base_off, bpr, _span_end) = type11_sample_window(m, mw, mh, format)?;
     Some((base_off, bpr, format))
-}
-
-/// Copy a resident target straight into the guest's pages, with the frame never
-/// existing on the host.
-///
-/// # What this is for
-///
-/// The copying rail this replaces moves the frame twice after the GPU has
-/// already written it once: the resident is read into a `HOST_VISIBLE` staging
-/// buffer, and the CPU then scatters that buffer into guest RAM row by row.
-/// `readback_split` prices the pair at 0.83 ms of staging map plus 2.68 ms of
-/// guest-page write inside a 6.9 ms flush, and the flush rail is 69% of the
-/// drain worker's second. Making the guest's own pages the copy's destination
-/// leaves only the copy that always had to happen.
-///
-/// # What it still owes
-///
-/// Everything [`write_bgra8_inner`] does *besides* moving bytes, because those
-/// obligations are about the guest's pages having changed and not about who
-/// changed them. In particular the guest-write witness
-/// ([`DeviceState::note_host_wrote_mapping`]) and the page footprint: a rail that
-/// lands frames without recording that it did makes
-/// [`crate::runtime::gather_witness`] attribute its own writes to the guest, and
-/// the type-11 resident rung above it then refuses residents and gathers whole
-/// surfaces per bind. That failure is measured and it costs more than this rail
-/// saves.
-///
-/// # Errors
-///
-/// Every decline is a routing answer: the caller still owes the frame and takes
-/// the copying rail. `Ok(())` means the pixels are in the guest's pages and the
-/// GPU has finished writing them.
-#[cfg(feature = "backend-vulkan")]
-pub fn write_bgra8_from_resident_gpu<M: HostMemory + HostOps>(
-    state: &mut DeviceState,
-    host: &mut M,
-    mapping_id: u32,
-    identity: &crate::backend::vulkan::engine::TargetIdentity,
-    width: u32,
-    height: u32,
-) -> Result<u64, GpuWritebackDecline> {
-    // This rail's own window, and the reason the licence does not resolve one:
-    // a Store's destination *is* the surface, so a frame whose extent is not
-    // the mapping's latched geometry is a frame for a mapping that has moved
-    // under it. That is a scanout rule and not a property of a type-11
-    // destination — a compute dispatch writing a sub-rectangle is ordinary — so
-    // it is asked here, by the caller it belongs to.
-    if !scanout_extent_ok(width, height) {
-        return Err(GpuWritebackDecline::NotWritable);
-    }
-    let Some(m) = state.mappings.get(&mapping_id) else {
-        return Err(GpuWritebackDecline::NotWritable);
-    };
-    if !m.mapped || m.page_entries.is_empty() {
-        return Err(GpuWritebackDecline::NotWritable);
-    }
-    let (mw, mh, format) = mapping_write_geometry(m, width, height);
-    if mw != width || mh != height {
-        return Err(GpuWritebackDecline::GeometryMoved {
-            latched_width: mw,
-            latched_height: mh,
-            frame_width: width,
-            frame_height: height,
-        });
-    }
-    let Some((base_off, bpr, span_end)) = type11_sample_window(m, mw, mh, format) else {
-        return Err(GpuWritebackDecline::WindowUnresolved {
-            width: mw,
-            height: mh,
-            format,
-        });
-    };
-    let licence = licence_type11_surface(
-        state,
-        host,
-        identity.resident_format(),
-        &Type11SurfaceDestination {
-            mapping_id,
-            base_off,
-            bpr,
-            span_end,
-            width: mw,
-            height: mh,
-            format,
-        },
-    )?;
-    crate::backend::vulkan::engine::copy_target_to_guest_pages(
-        identity,
-        &licence.target,
-        &licence.gpas,
-    )
-    .map_err(|inner| GpuWritebackDecline::Engine { inner })?;
-    note_type11_landed(state, mapping_id, licence.base_off, licence.span_end);
-    Ok(licence.span_end - licence.base_off)
-}
-
-/// A window of a type-11 surface mapping the GPU is asked to write.
-///
-/// The type-11 counterpart of
-/// [`crate::runtime::render_writeback::vulkan::GvaPlaneDestination`], and it exists for
-/// the same reason: the licence must not resolve its own destination. Which
-/// window of which plane a copy lands in is the *caller's* knowledge, and the
-/// two callers here come by it differently — a render Store's destination is the
-/// whole surface, while a compute bind resolves a window at stage time, which
-/// may be a sub-rectangle and, for a type-5 view, names its IOSurface plane on
-/// the wire.
-///
-/// Resolving it inside the licence instead served the Store and silently
-/// mis-served everything else. It refused every sub-rectangle — 15 of the 19
-/// remaining compute readbacks on a driven macos-13 boot, all
-/// [`GpuWritebackDecline::GeometryMoved`] — and behind that refusal sat a worse
-/// failure it was hiding: [`type11_sample_window`] takes no plane index and
-/// matches by geometry, so a type-5 bind's frame would have landed in whichever
-/// plane happened to share its dimensions. [`resident_gpu_plane`]'s doc states
-/// the cost of exactly that disagreement, which is that there is no error — the
-/// frame lands in the wrong plane of the right surface and the symptom is the
-/// next plane's pixels.
-///
-/// `format` is what the guest will read these bytes back as, and must be the
-/// format the window was resolved against; the licence derives the bytes per
-/// texel from it rather than taking one, so there is no second answer to carry.
-#[cfg(feature = "backend-vulkan")]
-pub(crate) struct Type11SurfaceDestination {
-    pub mapping_id: u32,
-    /// Byte offset of the window's first texel within the mapping.
-    pub base_off: u64,
-    /// Bytes per row of the surface, which is not `width * bpp`.
-    pub bpr: u32,
-    /// One past the last byte the window may touch.
-    pub span_end: u64,
-    pub width: u32,
-    pub height: u32,
-    pub format: u16,
-}
-
-/// A licensed direct-to-guest-pages destination over a type-11 surface mapping.
-///
-/// The type-11 counterpart of
-/// [`crate::runtime::render_writeback::vulkan::GvaPlaneLicence`], and it exists for the
-/// same reason: the two rails that write a guest surface from the GPU — a render
-/// Store and a compute storage-image output — differ only in the image they copy
-/// *from* and in which command buffer records the copy. Everything about the
-/// destination is this, so it is derived once and the second rail cannot spell
-/// any of it differently.
-///
-/// `base_off` and `span_end` are carried because the landing note needs them and
-/// deriving them a second time would be a second answer to a question the
-/// licence already asked. See [`note_type11_landed`].
-#[cfg(feature = "backend-vulkan")]
-pub(crate) struct Type11SurfaceLicence {
-    pub target: crate::backend::vulkan::engine::GuestPageTarget,
-    /// The pages walked, in the surface's own order — what the copy is licensed
-    /// over and what every witness on this path is armed against.
-    pub gpas: Vec<u64>,
-    /// The window within the mapping the copy names, and nothing past it.
-    pub base_off: u64,
-    pub span_end: u64,
-}
-
-/// Licence a caller's type-11 surface window as a destination the GPU may copy
-/// into, or give the typed reason it may not.
-///
-/// Takes the window rather than resolving one — see
-/// [`Type11SurfaceDestination`] for why that division is where it is. What is
-/// shared between the two rails, and therefore lives here, is the format rule,
-/// the page-list plan, the page walk, the guest-RAM references and the two
-/// guest-write witnesses.
-///
-/// `held` is the format the source image actually holds. **A copy converts
-/// nothing**, so a source whose format is not the one the guest will read these
-/// bytes back as must be refused here rather than landed.
-///
-/// # Why this asks about the format and the render caller used not to
-///
-/// On the render rail the question is very nearly a tautology — a target's
-/// identity takes its format from [`mapping_store_format`], the same function
-/// that caller resolved its own window through — and the comparison that can
-/// actually fail is made downstream by `copy_target_to_guest_pages`, against the
-/// resident *image's* own format, which a mapping may have redeclared
-/// underneath. Both of those are still true and that check is still there.
-///
-/// The compute rail has no such downstream. Its dispatch writes the guest's
-/// pages itself, so there is no second pair to compare and this is the only
-/// place the question can be asked. A storage image takes its format from the
-/// specialized SPIR-V texel format, which has no reason to match the format the
-/// bind staged its window against — banded at 3 of 35 type-11 windows on a
-/// driven macos-13 boot — so on that rail it is not a healthy zero either.
-///
-/// Asking it for both callers rather than for the one that needs it keeps the
-/// rule in the licence, where a third writer of a guest surface would meet it
-/// without knowing to look.
-#[cfg(feature = "backend-vulkan")]
-pub(crate) fn licence_type11_surface<M: HostMemory + HostOps>(
-    state: &mut DeviceState,
-    host: &mut M,
-    held: ash::vk::Format,
-    dst: &Type11SurfaceDestination,
-) -> Result<Type11SurfaceLicence, GpuWritebackDecline> {
-    let &Type11SurfaceDestination {
-        mapping_id,
-        base_off,
-        bpr,
-        span_end,
-        width: mw,
-        height: mh,
-        format,
-    } = dst;
-    // The mapping still has to be here and still have to be backed — the caller
-    // resolved a window against it, and may have done so before this call.
-    let Some(m) = state.mappings.get(&mapping_id) else {
-        return Err(GpuWritebackDecline::NotWritable);
-    };
-    if !m.mapped || m.page_entries.is_empty() {
-        return Err(GpuWritebackDecline::NotWritable);
-    }
-    // An image→buffer copy moves bytes and converts nothing, so this rail can
-    // only serve a window whose declared format is a host texel verbatim. A
-    // compressed or planar declaration is not, and the copying rail's row
-    // converter is the only thing that can land it.
-    //
-    // Asked of every rail that creates images rather than of the render Store's
-    // table alone: this licence serves a compute storage output as well as a
-    // Store, and a storage image is a thing the guest neither renders into nor
-    // samples. See
-    // [`crate::backend::vulkan::translate::pixel::verbatim_texel`].
-    let Some((dst_format, texel)) =
-        crate::backend::vulkan::translate::pixel::verbatim_texel(format)
-    else {
-        return Err(GpuWritebackDecline::FormatNeedsConversion { format });
-    };
-    // And that the source holds exactly it. See this function's doc for why the
-    // render caller's own downstream check stays where it is: the two compare
-    // different pairs, and this is the only one the compute rail has.
-    if held != dst_format {
-        return Err(GpuWritebackDecline::ResidentFormatMismatch {
-            held,
-            want: dst_format,
-        });
-    }
-    let shared_backing = if host.map_pages_stable() {
-        mapper::ensure_contig_view(state, host, mapping_id).map(|(ptr, len)| {
-            crate::backend::vulkan::engine::GuestTargetBacking {
-                allocation_host_ptr: ptr,
-                allocation_len: len as u64,
-                plane_offset: base_off,
-                row_pitch: u64::from(bpr),
-            }
-        })
-    } else {
-        None
-    };
-    // One live window per physical format is enough to answer the remaining
-    // device-level question: whether the driver's actual linear layout and
-    // memory requirements agree with a guest plane on this host. The packed
-    // view is the allocation a direct image retains. The report remains useful
-    // because creation declines are per target and this gives the full binding
-    // equation once per physical format.
-    let probe_key = dst_format.as_raw() as u32 as u64;
-    if crate::observe::first_sight("vk_linear_target_window_probe", probe_key) {
-        if let Some(backing) = shared_backing {
-            crate::backend::vulkan::engine::probe_guest_backed_target(
-                backing.allocation_host_ptr,
-                backing.allocation_len,
-                backing.plane_offset,
-                backing.row_pitch,
-                mw,
-                mh,
-                dst_format,
-            );
-        } else {
-            crate::observe::off(format!(
-                "vk_linear_target_window verdict=no_packed_alias format={dst_format:?} {mw}x{mh} plane_offset={base_off} guest_row_pitch={bpr}"
-            ));
-        }
-    }
-    // No settle here, and the twin rail is why. `render_writeback::vulkan::store_gva_frame`
-    // does exactly this for a GVA-addressed destination — vouch, resolve runs,
-    // submit a buffer copy — and takes no settle at all, because nothing between
-    // here and the submit reads the pixel bytes a pending writeback would land
-    // in: the page list comes from `page_entries` (device state), the vouch
-    // walks the guest's page tables, and `references_for_runs` resolves host
-    // pointers. The copy itself is a GPU command on the same single queue as
-    // any outstanding writeback, so queue order already puts the older write
-    // ahead of this one and a CPU fence buys an ordering that holds without it.
-    // That is the same argument `try_linear_sample_zero_copy` states for its own
-    // gather.
-    //
-    // What used to stand here was the deferred rail's flush-on-access, whose
-    // justification was that landing a pending window "can invalidate the
-    // mapping". There are no windows: a type-11 render Store lands its frame at
-    // the Store (see `render_writeback`'s module doc). Measured at **5 204
-    // settles and 7.29 s blocked** on a driven Safari-drag boot — 42 % of every
-    // wait in the device — for an ordering the queue already had.
-    //
-    // Nothing below can land a frame on a host whose GPU cannot import guest
-    // RAM, so the walks below are skipped rather than run and discarded.
-    //
-    // Not a second gate — it is the *same* gate, asked earlier.
-    // `references_for_runs` below opens with the identical question and would
-    // decline these same pages a few hundred microseconds later; this only
-    // declines sooner, and on a pathway where the answer never changes that
-    // saves a page-table walk per flush for the life of the process.
-    //
-    // Asked through `standing_refusal` rather than by re-reading the
-    // granularity latch, because the latch is only one of the four things that
-    // refuse here — a shim that cannot say where guest RAM lives and a machine
-    // whose every span failed the bound both leave a granularity published, and
-    // used to walk the whole page list before finding that out.
-    if let Some(refusal) = crate::runtime::guest_ram_map::standing_refusal(host) {
-        return Err(GpuWritebackDecline::GuestRefRefused { refusal });
-    }
-    // Timed on its own because it is the largest `O(pages)` step left and its
-    // fix is not the other one's. `vouch_for_write` re-walks every page of the
-    // mapping through the guest's page table — the check that licenses writing
-    // to them at all — and until the host copies were removed that cost was
-    // hidden inside a millisecond of memcpy.
-    use crate::runtime::drain::{note_readback_phase, ReadbackPhase};
-    let vouch_started = std::time::Instant::now();
-    let vouched = vouch_for_write(state, host, mapping_id, "gpu_writeback");
-    note_readback_phase(
-        ReadbackPhase::Vouch,
-        vouch_started.elapsed().as_micros() as u64,
-    );
-    if vouched.is_none() {
-        return Err(GpuWritebackDecline::PagesNotOurs);
-    }
-    let resolve_started = std::time::Instant::now();
-    let page_size = state.page_size();
-    let page_shift = state.page_shift;
-    let Some(m) = state.mappings.get(&mapping_id) else {
-        return Err(GpuWritebackDecline::NotWritable);
-    };
-    let plan = plan_guest_window(
-        m.page_entries.len(),
-        page_size,
-        base_off,
-        span_end,
-        bpr,
-        mw,
-        texel,
-    )?;
-    let mut gpas = Vec::with_capacity(plan.pages());
-    for (i, &entry) in m.page_entries[plan.first_page..=plan.last_page]
-        .iter()
-        .enumerate()
-    {
-        let Some(gpa) = crate::contract::iosurface_pages::entry_gpa_shift(entry, page_shift) else {
-            return Err(GpuWritebackDecline::PageUnbacked {
-                index: plan.first_page + i,
-            });
-        };
-        gpas.push(gpa);
-    }
-    // The extent this copy names, and nothing past it. Padding after the final
-    // row belongs to the surface's plane but is not texels this call was given,
-    // and the copying rail does not write it either — a request that included
-    // it would let the two rails land different guest memory for one frame.
-    let pitch = u64::from(plan.row_length_texels.max(mw)) * u64::from(texel);
-    let extent = u64::from(mh.saturating_sub(1)) * pitch + u64::from(mw) * u64::from(texel);
-    // Every run, not one range. The guest backs a surface in 16 KiB granules
-    // with no relation to each other, so a full-screen window is ~507 stretches
-    // and asking for a single contiguous reference refused every 1080p flush of
-    // a driven boot.
-    let runs = crate::runtime::guest_ram_map::references_for_runs(
-        host,
-        &gpas,
-        page_size,
-        plan.in_page,
-        extent,
-    )
-    .map_err(|refusal| GpuWritebackDecline::GuestRefRefused { refusal })?;
-    let target = crate::backend::vulkan::engine::GuestPageTarget {
-        runs,
-        row_length_texels: plan.row_length_texels,
-        width: mw,
-        height: mh,
-        // The guest's own declaration for this plane, which is what the guest
-        // will read these bytes back as. Every byte offset planned above came
-        // from its width, and the engine refuses the copy outright if the
-        // resident does not hold exactly this.
-        format: dst_format,
-        shared_backing,
-    };
-    // Both witnesses before the copy rather than after it, matching
-    // `contig_for_write`: a refused write costs a spurious bump, which makes a
-    // reader re-read bytes that did not change, while the opposite error hands
-    // out a stale copy as fresh.
-    //
-    // Marked over `[base_off, span_end)` — the extent the copy names — rather
-    // than from zero, because unlike the contig view this rail knows its own
-    // rows and has no pointer whose coverage it is restating.
-    mapper::note_mapping_write_footprint(state, mapping_id, base_off, span_end - base_off);
-    state.note_host_wrote_mapping(mapping_id);
-    note_readback_phase(
-        ReadbackPhase::Resolve,
-        resolve_started.elapsed().as_micros() as u64,
-    );
-    Ok(Type11SurfaceLicence {
-        target,
-        gpas,
-        base_off,
-        span_end,
-    })
-}
-
-/// What a landed type-11 GPU write owes the rest of the device.
-///
-/// Called once the copy is *issued*, not once it has completed, and by both
-/// rails that issue one — the render Store, which submits and waits, and the
-/// compute storage-image output, whose copy rides its dispatch's own command
-/// buffer and lands at the fence. Neither leaves a host copy of the frame, so
-/// nothing on the host may go on naming one, and that is true from the moment
-/// the copy is queued rather than from the moment it retires.
-///
-/// Every one of these errs in the same direction on purpose: a cache forgotten
-/// early costs a re-read of bytes that are about to change anyway, while one
-/// forgotten late hands out a stale frame as fresh. The same argument the
-/// witnesses in [`licence_type11_surface`] are armed on.
-#[cfg(feature = "backend-vulkan")]
-pub(crate) fn note_type11_landed(
-    state: &mut DeviceState,
-    mapping_id: u32,
-    base_off: u64,
-    span_end: u64,
-) {
-    state.invalidate_storage_residency_window(mapping_id, base_off, span_end);
-    let _ = state.mark_mapping_written(mapping_id);
-    // The guest's pages are now the only place this frame exists, and the
-    // surface cache's entry, if any, is a previous flush's bytes. Same reason
-    // `write_bgra8_uncached` invalidates rather than publishes.
-    crate::runtime::surface_cache::forget(state, mapping_id);
-}
-
-/// Publish a Store from an attachment already backed by this mapping.
-///
-/// Import admission retained the mapping's bounded allocation and physical-page
-/// footprint in the resident. Synchronization therefore names that resident;
-/// it does not reconstruct a `GuestPageTarget`, re-walk the page table, or
-/// reacquire one guest reference per page.  Those operations describe a copy
-/// destination, and this path has no copy destination—the attachment is the
-/// guest allocation.
-#[cfg(feature = "backend-vulkan")]
-pub fn synchronize_guest_backed_resident(
-    state: &mut DeviceState,
-    mapping_id: u32,
-    identity: &crate::backend::vulkan::engine::TargetIdentity,
-    width: u32,
-    height: u32,
-    guest_store_recorded: bool,
-    guest_store_footprint: Option<crate::runtime::guest_ram::GuestPageFootprint>,
-) -> Result<u64, GpuWritebackDecline> {
-    if !scanout_extent_ok(width, height) {
-        return Err(GpuWritebackDecline::NotWritable);
-    }
-    let Some(m) = state.mappings.get(&mapping_id) else {
-        return Err(GpuWritebackDecline::NotWritable);
-    };
-    if !m.mapped || m.page_entries.is_empty() {
-        return Err(GpuWritebackDecline::NotWritable);
-    }
-    let (mw, mh, format) = mapping_write_geometry(m, width, height);
-    if mw != width || mh != height {
-        return Err(GpuWritebackDecline::GeometryMoved {
-            latched_width: mw,
-            latched_height: mh,
-            frame_width: width,
-            frame_height: height,
-        });
-    }
-    let Some((base_off, _bpr, span_end)) = type11_sample_window(m, mw, mh, format) else {
-        return Err(GpuWritebackDecline::WindowUnresolved {
-            width: mw,
-            height: mh,
-            format,
-        });
-    };
-    let footprint = if guest_store_needs_separate_sync(guest_store_recorded) {
-        crate::backend::vulkan::engine::synchronize_guest_backed_target(identity)
-            .map_err(|inner| GpuWritebackDecline::Engine { inner })?
-    } else {
-        guest_store_footprint.ok_or(GpuWritebackDecline::Engine {
-            inner: crate::backend::vulkan::engine::DrawError::GuestPageWrite(
-                crate::backend::vulkan::engine::GuestWriteDecline::NoSharedBacking,
-            ),
-        })?
-    };
-
-    mapper::note_physical_page_write_footprint(&footprint, base_off, span_end - base_off);
-    state.host_writes.note_footprint(&footprint);
-    state.invalidate_storage_residency_window(mapping_id, base_off, span_end);
-    let _ = state.mark_mapping_written(mapping_id);
-    crate::runtime::surface_cache::forget(state, mapping_id);
-    Ok(span_end - base_off)
-}
-
-#[cfg(feature = "backend-vulkan")]
-fn guest_store_needs_separate_sync(recorded_in_draw: bool) -> bool {
-    !recorded_in_draw
 }
 
 /// The geometry and pixel format a writeback to this mapping must land in.
